changed the walrus operator location to resolve circular import
