+++ src/astx/__init__.py
@@ -2,7 +2,6 @@
 """ASTx."""
 
 from importlib import metadata as importlib_metadata
-from astx.types.operators import WalrusOp
 
 from astx import (
     base,
@@ -92,6 +91,7 @@ from astx.modifiers import (
     ScopeKind,
     VisibilityKind,
 )
+from astx.operators import WalrusOp
 from astx.packages import (
     AliasExpr,
     ImportExpr,
@@ -155,125 +155,125 @@ def get_version() -> str:
 
 
 __all__ = [
+    "AST",
+    "ASTKind",
     "AliasExpr",
     "Argument",
     "Arguments",
-    "AST",
-    "ASTKind",
-    "base",
     "BinaryOp",
     "Block",
-    "blocks",
     "Boolean",
-    "callables",
+    "ClassDeclStmt",
+    "ClassDefStmt",
+    "Complex",
+    "Complex32",
+    "Complex64",
     "DataType",
     "DataTypeOps",
-    "datatypes",
+    "Date",
+    "DateTime",
+    "EnumDeclStmt",
     "Expr",
     "ExprType",
     "Float16",
     "Float32",
     "Float64",
-    "LiteralFloat16",
-    "LiteralFloat32",
-    "LiteralFloat64",
     "Floating",
-    "flows",
-    "ForCountLoopStmt",
     "ForCountLoopExpr",
-    "ForRangeLoopStmt",
+    "ForCountLoopStmt",
     "ForRangeLoopExpr",
+    "ForRangeLoopStmt",
     "Function",
     "FunctionCall",
     "FunctionPrototype",
     "FunctionReturn",
-    "get_version",
-    "IfStmt",
     "IfExpr",
-    "ImportFromExpr",
+    "IfStmt",
     "ImportExpr",
-    "ImportStmt",
+    "ImportFromExpr",
     "ImportFromStmt",
+    "ImportStmt",
     "InlineVariableDeclaration",
+    "Int8",
     "Int16",
     "Int32",
     "Int64",
-    "Int8",
     "Integer",
-    "UnsignedInteger",
-    "UInt8",
-    "UInt16",
-    "UInt32",
-    "UInt64",
-    "UInt128",
     "LambdaExpr",
     "Literal",
-    "LiteralNone",
     "LiteralBoolean",
+    "LiteralComplex",
+    "LiteralComplex32",
+    "LiteralComplex64",
+    "LiteralDate",
+    "LiteralDateTime",
+    "LiteralFloat16",
+    "LiteralFloat32",
+    "LiteralFloat64",
     "LiteralInt8",
     "LiteralInt16",
     "LiteralInt32",
     "LiteralInt64",
     "LiteralInt128",
+    "LiteralNone",
+    "LiteralString",
+    "LiteralTime",
+    "LiteralTimestamp",
     "LiteralUInt8",
     "LiteralUInt16",
     "LiteralUInt32",
     "LiteralUInt64",
     "LiteralUInt128",
-    "mixes",
+    "LiteralUTF8Char",
+    "LiteralUTF8String",
     "Module",
     "MutabilityKind",
     "NamedExpr",
     "Number",
     "OperatorType",
-    "packages",
     "Package",
     "Program",
     "ScopeKind",
     "SignedInteger",
     "SourceLocation",
     "StatementType",
-    "SubscriptExpr",
-    "symbol_table",
-    "Target",
-    "TypeCastExpr",
-    "UnaryOp",
-    "Undefined",
-    "VariableAssignment",
-    "VariableDeclaration",
-    "Variable",
-    "variables",
-    "VisibilityKind",
-    "WalrusOp",
-    "WhileStmt",
-    "WhileExpr",
-    "Complex",
-    "Complex32",
-    "Complex64",
-    "ClassDefStmt",
-    "ClassDeclStmt",
-    "EnumDeclStmt",
+    "String",
     "StructDeclStmt",
     "StructDefStmt",
-    "LiteralComplex",
-    "LiteralComplex32",
-    "LiteralComplex64",
-    "LiteralUTF8Char",
-    "LiteralString",
-    "LiteralUTF8String",
-    "UTF8Char",
-    "String",
-    "UTF8String",
-    "Date",
-    "DateTime",
-    "Timestamp",
+    "SubscriptExpr",
+    "Target",
     "Time",
-    "LiteralDate",
-    "LiteralDateTime",
-    "LiteralTimestamp",
-    "LiteralTime",
-    "types",
+    "Timestamp",
+    "TypeCastExpr",
+    "UInt8",
+    "UInt16",
+    "UInt32",
+    "UInt64",
+    "UInt128",
+    "UTF8Char",
+    "UTF8String",
+    "UnaryOp",
+    "Undefined",
+    "UnsignedInteger",
+    "Variable",
+    "VariableAssignment",
+    "VariableDeclaration",
+    "VisibilityKind",
+    "WalrusOp",
+    "WhileExpr",
+    "WhileStmt",
+    "base",
+    "blocks",
+    "callables",
+    "datatypes",
+    "flows",
+    "get_version",
     "literals",
+    "mixes",
+    "packages",
+    "symbol_table",
+    "types",
+    "variables",
 ]
